% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotRegulatoryNetwork}
\alias{plotRegulatoryNetwork}
\title{Plot regulatory networks}
\usage{
plotRegulatoryNetwork(regnet, geneColor = "#b0b0b0",
  mirnaColor = "SkyBlue2", nodeSize = 5, arrowSize = 0.2, showLabel = F)
}
\arguments{
\item{regnet}{A regulatory network}

\item{geneColor}{the color used to plot genes}

\item{mirnaColor}{the color used to plot mirnas}

\item{nodeSize}{the size of the nodes}

\item{arrowSize}{the size of the edges arrows}

\item{showLabel}{logical, should node labels be displayed?}
}
\description{
Plots a regulatory network
}

