% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RegulatoryNetwork}
\alias{RegulatoryNetwork}
\title{Generation of regulatory networks}
\usage{
RegulatoryNetwork(genes = 200, mirnas = 100, genes.in_degree = 2.2,
  genes.out_degree = 2.2, mirnas.in_degree = 2.2, mirnas.out_degree = 0.1,
  clustering_coefficient = 0.2, max_regulators = 5, max_size = c(5, 8, 5,
  5), pool_size = 30, repression = 0.35, control_genes = 5,
  cooperation = 0.5, progress = c("none", "text", "gui"))
}
\arguments{
\item{genes}{number of genes}

\item{mirnas}{number of mirnas}

\item{genes.in_degree}{a numeric value that specifies the scale parameter of a power-law degree distribution for the in degree of genes, or a numeric vector specifying a custom degree distribution}

\item{genes.out_degree}{a numeric value that specifies the scale parameter of a power-law degree distribution for the out degree of genes, or a numeric vector specifying a custom degree distribution}

\item{mirnas.in_degree}{a numeric value that specifies the parameter lambda of an exponential degree distribution for the in degree of mirnas, or a numeric vector specifying a custom degree distribution}

\item{mirnas.out_degree}{a numeric value that specifies the parameter lambda of an exponential degree distribution for the out degree of mirnas, or a numeric vector specifying a custom degree distribution}

\item{clustering_coefficient}{target clustering coefficient of the network}

\item{max_regulators}{the maximum number of regulators that a node can have}

\item{max_size}{vector of integers that specifies the size of each motif, if not sure leave it as default}

\item{pool_size}{size of the pool of generated motifs for each step of the procedure}

\item{repression}{probability that regulator represses one of its regulated nodes}

\item{control_genes}{number of signalling genes controlling the state of the network durign simulation}

\item{cooperation}{probability that two or more regulating nodes cooperate}

\item{progress}{the type of progress bar to be shown}
}
\description{
Generation of a transcriptional and post-trascriptional regulatory network
}
\details{

}

