% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateExperiment}
\alias{simulateExperiment}
\title{Simulation of a biological experiment}
\usage{
simulateExperiment(regnet, INITIAL, SIGNAL, times = 1:100, sigmaEta = 0.5,
  sigmaEps = 3000)
}
\arguments{
\item{regnet}{a regulatory network to be simulated}

\item{INITIAL}{a matrix containing initial values for each node of the network for each sample to be simulated}

\item{SIGNAL}{a list containing a list of functions for each sample to be simulated}

\item{times}{timescale in which simulation will happen}

\item{sigmaEta}{standard deviation parameter of the model of technical noise}

\item{sigmaEps}{standard deviation parameter of the model of technical noise}
}
\description{
Simulates a whole multi-view microarray experiment.
}

